+++ packages/at_secondary_server/test/monitor_verb_test.dart
@@ -11,7 +11,6 @@ import 'package:at_secondary/src/verb/handler/enroll_verb_handler.dart';
 import 'package:at_secondary/src/verb/handler/monitor_verb_handler.dart';
 import 'package:at_secondary/src/verb/handler/otp_verb_handler.dart';
 import 'package:at_server_spec/at_server_spec.dart';
-import 'package:test/expect.dart';
 import 'package:test/test.dart';
 import 'package:uuid/uuid.dart';
 
@@ -351,7 +350,7 @@ void main() {
         {required bool autoApprove}) async {
       OtpVerbHandler otpVH = OtpVerbHandler(secondaryKeyStore);
       String otp = otpVH.generateOTP();
-      await otpVH.saveOTP(otp, 5000);
+      await otpVH.savePasscode(otp, ttl: 5000, isSpp: false);
 
       EnrollVerbHandler enrollVerbHandler =
           EnrollVerbHandler(secondaryKeyStore);
@@ -361,7 +360,7 @@ void main() {
           ',"deviceName":"$deviceName"'
           ',"namespaces":${jsonEncode(namespaces)}'
           ',"apkamPublicKey":"dummy_apkam_public_key"'
-          ',"encryptedAPKAMSymmetricKey":"dummy_encrypted_apkam_symm_key"'
+          ',"encryptedAPKAMSymmetricKey":"dummy_encrypted_apkam_symmetric_key"'
           '}';
       HashMap<String, String?> enrollmentRequestVerbParams =
           getVerbParam(VerbSyntax.enroll, enrollmentRequest);
@@ -480,9 +479,9 @@ void main() {
       final valueJson = jsonDecode(notificationJson['value']);
       //TODO remove encryptedApkamSymmetricKey in the future
       expect(valueJson['encryptedApkamSymmetricKey'],
-          'dummy_encrypted_apkam_symm_key');
+          'dummy_encrypted_apkam_symmetric_key');
       expect(valueJson['encryptedAPKAMSymmetricKey'],
-          'dummy_encrypted_apkam_symm_key');
+          'dummy_encrypted_apkam_symmetric_key');
       expect(valueJson['appName'], 'mvt_app_2');
       expect(valueJson['deviceName'], 'mvt_dev_2');
       expect(valueJson['namespace'], equals({'app_2_namespace': 'rw'}));