+++ packages/at_secondary_server/test/otp_verb_test.dart
@@ -90,7 +90,7 @@ void main() {
       OtpVerbHandler otpVerbHandler = OtpVerbHandler(secondaryKeyStore);
       await otpVerbHandler.processVerb(response, verbParams, inboundConnection);
       String? otp = response.data;
-      expect(await otpVerbHandler.isOTPValid(otp), true);
+      expect(await otpVerbHandler.isPasscodeValid(otp), true);
     });
 
     test('A test to verify otp:get with TTL set expires after the TTL is met',
@@ -103,7 +103,7 @@ void main() {
       await otpVerbHandler.processVerb(response, verbParams, inboundConnection);
       String? otp = response.data;
       await Future.delayed(Duration(seconds: 1));
-      expect(await otpVerbHandler.isOTPValid(otp), false);
+      expect(await otpVerbHandler.isPasscodeValid(otp), false);
     });
     tearDown(() async => await verbTestsTearDown());
   });
@@ -113,7 +113,7 @@ void main() {
       await verbTestsSetUp();
     });
     test(
-        'A test to verify UnAuthorizedException is thrown when opt verb is executed on an unauthenticated conn',
+        'A test to verify UnAuthorizedException is thrown when opt:get is executed on an unauthenticated conn',
         () {
       Response response = Response();
       HashMap<String, String?> verbParams =
@@ -124,8 +124,9 @@ void main() {
               response, verbParams, inboundConnection),
           throwsA(predicate((e) =>
               e is UnAuthenticatedException &&
-              e.message == 'otp:get requires authenticated connection')));
+              e.message == 'otp: requires authenticated connection')));
     });
+
     tearDown(() async => await verbTestsTearDown());
   });
 
@@ -141,7 +142,7 @@ void main() {
       inboundConnection.metaData.isAuthenticated = true;
       OtpVerbHandler otpVerbHandler = OtpVerbHandler(secondaryKeyStore);
       await otpVerbHandler.processVerb(response, verbParams, inboundConnection);
-      expect(await otpVerbHandler.isOTPValid(response.data), true);
+      expect(await otpVerbHandler.isPasscodeValid(response.data), true);
     });
 
     test('A test to verify otp:validate returns invalid when OTP is expired',
@@ -154,7 +155,7 @@ void main() {
       await otpVerbHandler.processVerb(response, verbParams, inboundConnection);
       String? otp = response.data;
       await Future.delayed(Duration(milliseconds: 2));
-      expect(await otpVerbHandler.isOTPValid(otp), false);
+      expect(await otpVerbHandler.isPasscodeValid(otp), false);
     });
 
     test('A test to verify default otp expiry not overwritten', () async {
@@ -182,7 +183,7 @@ void main() {
         () async {
       String otp = 'ABC123';
       OtpVerbHandler otpVerbHandler = OtpVerbHandler(secondaryKeyStore);
-      expect(await otpVerbHandler.isOTPValid(otp), false);
+      expect(await otpVerbHandler.isPasscodeValid(otp), false);
     });
 
     test('A test to verify otp is removed from the keystore after use',
@@ -194,8 +195,8 @@ void main() {
       OtpVerbHandler otpVerbHandler = OtpVerbHandler(secondaryKeyStore);
       await otpVerbHandler.processVerb(response, verbParams, inboundConnection);
       String? otp = response.data;
-      expect(await otpVerbHandler.isOTPValid(otp), true);
-      expect(await otpVerbHandler.isOTPValid(otp), false);
+      expect(await otpVerbHandler.isPasscodeValid(otp), true);
+      expect(await otpVerbHandler.isPasscodeValid(otp), false);
     });
 
     test('validate backwards compatability with legacy otp key', () async {
@@ -209,7 +210,7 @@ void main() {
       await secondaryKeyStore.put(otpLegacyKey, value);
 
       OtpVerbHandler otpVerbHandler = OtpVerbHandler(secondaryKeyStore);
-      expect(await otpVerbHandler.isOTPValid(testOtp), true);
+      expect(await otpVerbHandler.isPasscodeValid(testOtp), true);
     });
     tearDown(() async => await verbTestsTearDown());
   });
@@ -227,6 +228,7 @@ void main() {
       await secondaryKeyStore.put(
           enrollmentKey, AtData()..data = jsonEncode(enrollDataStoreValue));
     });
+
     test('A test to set a pass code and verify isOTPValid returns true',
         () async {
       String passcode = 'abc123';
@@ -238,10 +240,65 @@ void main() {
           enrollmentId;
       OtpVerbHandler otpVerbHandler = OtpVerbHandler(secondaryKeyStore);
       await otpVerbHandler.processVerb(response, verbParams, inboundConnection);
-      expect(await otpVerbHandler.isOTPValid(passcode), true);
+      expect(await otpVerbHandler.isPasscodeValid(passcode), true);
       // Adding expect again to ensure the Semi-permanent passcodes are not deleted
       // after one time use.
-      expect(await otpVerbHandler.isOTPValid(passcode), true);
+      expect(await otpVerbHandler.isPasscodeValid(passcode), true);
+    });
+
+    test('set spp with a ttl, check isOTPValid before ttl expires', () async {
+      String passcode = 'SppWithTtl50';
+      Response response = Response();
+      HashMap<String, String?> verbParams =
+          getVerbParam(VerbSyntax.otp, 'otp:put:$passcode:ttl:50');
+      inboundConnection.metaData.isAuthenticated = true;
+      (inboundConnection.metaData as InboundConnectionMetadata).enrollmentId =
+          enrollmentId;
+      OtpVerbHandler otpVerbHandler = OtpVerbHandler(secondaryKeyStore);
+      await otpVerbHandler.processVerb(response, verbParams, inboundConnection);
+      expect(await otpVerbHandler.isPasscodeValid(passcode), true);
+      // Adding expect again to ensure the Semi-permanent passcodes are not deleted
+      // after one time use.
+      await Future.delayed(Duration(milliseconds: 10));
+      expect(await otpVerbHandler.isPasscodeValid(passcode), true);
+    });
+
+    test('set spp with a ttl, check isOTPValid before and after ttl expires',
+        () async {
+      String passcode = 'SppWithTtl50';
+      Response response = Response();
+      HashMap<String, String?> verbParams =
+          getVerbParam(VerbSyntax.otp, 'otp:put:$passcode:ttl:50');
+      inboundConnection.metaData.isAuthenticated = true;
+      (inboundConnection.metaData as InboundConnectionMetadata).enrollmentId =
+          enrollmentId;
+      OtpVerbHandler otpVerbHandler = OtpVerbHandler(secondaryKeyStore);
+      await otpVerbHandler.processVerb(response, verbParams, inboundConnection);
+      expect(await otpVerbHandler.isPasscodeValid(passcode), true);
+      // Adding expect again to ensure the Semi-permanent passcodes are not deleted
+      // after one time use.
+      expect(await otpVerbHandler.isPasscodeValid(passcode), true);
+
+      await Future.delayed(Duration(milliseconds: 51));
+      expect(await otpVerbHandler.isPasscodeValid(passcode), false);
+    });
+
+    test('set spp without a ttl, verify its ttl has been set to -1', () async {
+      String passcode = 'SppWithoutTtl';
+      Response response = Response();
+      HashMap<String, String?> verbParams =
+          getVerbParam(VerbSyntax.otp, 'otp:put:$passcode');
+      inboundConnection.metaData.isAuthenticated = true;
+      (inboundConnection.metaData as InboundConnectionMetadata).enrollmentId =
+          enrollmentId;
+      OtpVerbHandler otpVerbHandler = OtpVerbHandler(secondaryKeyStore);
+      await otpVerbHandler.processVerb(response, verbParams, inboundConnection);
+      await Future.delayed(Duration(milliseconds: 2));
+      expect(await otpVerbHandler.isPasscodeValid(passcode), true);
+
+      String sppKey = OtpVerbHandler.passcodeKey(passcode, isSpp: true);
+      var atData = await secondaryKeyStore.get(sppKey);
+      expect(atData?.metaData?.ttl, -1);
     });
 
     test('A test to verify pass code can be updated', () async {
@@ -254,14 +311,14 @@ void main() {
           enrollmentId;
       OtpVerbHandler otpVerbHandler = OtpVerbHandler(secondaryKeyStore);
       await otpVerbHandler.processVerb(response, verbParams, inboundConnection);
-      expect(await otpVerbHandler.isOTPValid(passcode), true);
+      expect(await otpVerbHandler.isPasscodeValid(passcode), true);
       // Update the pass-code
       passcode = 'xyz987';
       response = Response();
       verbParams = getVerbParam(VerbSyntax.otp, 'otp:put:$passcode');
       otpVerbHandler = OtpVerbHandler(secondaryKeyStore);
       await otpVerbHandler.processVerb(response, verbParams, inboundConnection);
-      expect(await otpVerbHandler.isOTPValid(passcode), true);
+      expect(await otpVerbHandler.isPasscodeValid(passcode), true);
     });
 
     test('validate backwards compatability with legacy ssp key', () async {
@@ -272,7 +329,7 @@ void main() {
       await secondaryKeyStore.put(otpLegacyKey, value);
 
       OtpVerbHandler otpVerbHandler = OtpVerbHandler(secondaryKeyStore);
-      expect(await otpVerbHandler.isOTPValid(testOtp), true);
+      expect(await otpVerbHandler.isPasscodeValid(testOtp), true);
     });
     tearDown(() async => await verbTestsTearDown());
   });